Call precalculate_grid when new POIs are added.
When rows are insterted into pois_modified, a trigger is called. The trigger now checks which cells are influenced by the inserted geometry and calls precalculate grid on those cells. Performance is not good.
And to actually find the new features, closest_pois() would have to be modified to look for pois in the pois_userinput table.
Indices for walkability are not recalculated yet.